requirements.txt
spacy
python-telegram-bot
python-dotenv